Move example project into its own root-level crate
