+++ project-example/schema/example.schema
