--- spatialos-sdk/examples/project-example/.gitignore
@@ -1,2 +1,0 @@
-# Generated bindings to SpatialOS components, shouldn't be committed.
-generated_code.rs
