--- spatialos-sdk/examples/project-example/lib/mod.rs
@@ -1,5 +1,0 @@
-pub mod connection_handler;
-pub mod opt;
-
-pub use self::connection_handler::*;
-pub use self::opt::*;
