--- spatialos-sdk/examples/project-example/spatialos/.gitignore
@@ -1,3 +1,0 @@
-.improbable
-logs
-schema/bin
